--- server/routes/testutils/appSetup.ts
@@ -1,59 +1,0 @@
-import express, { Express } from 'express'
-import cookieSession from 'cookie-session'
-import createError from 'http-errors'
-import path from 'path'
-
-import routes from '../index'
-import nunjucksSetup from '../../utils/nunjucksSetup'
-import errorHandler from '../../errorHandler'
-import * as auth from '../../authentication/auth'
-import { Controllers } from '../../controllers'
-
-export const user = {
-  firstName: 'first',
-  lastName: 'last',
-  userId: 'id',
-  token: 'token',
-  username: 'user1',
-  displayName: 'First Last',
-  activeCaseLoadId: 'MDI',
-  authSource: 'NOMIS',
-}
-
-export const flashProvider = jest.fn()
-
-function appSetup(controllers: Controllers, production: boolean, userSupplier: () => Express.User): Express {
-  const app = express()
-
-  app.set('view engine', 'njk')
-
-  nunjucksSetup(app, path)
-  app.use(cookieSession({ keys: [''] }))
-  app.use((req, res, next) => {
-    req.user = userSupplier()
-    req.flash = flashProvider
-    res.locals = {}
-    res.locals.user = { ...req.user }
-    next()
-  })
-  app.use(express.json())
-  app.use(express.urlencoded({ extended: true }))
-  app.use(routes(controllers))
-  app.use((req, res, next) => next(createError(404, 'Not found')))
-  app.use(errorHandler(production))
-
-  return app
-}
-
-export function appWithAllRoutes({
-  production = false,
-  controllers = {},
-  userSupplier = () => user,
-}: {
-  production?: boolean
-  controllers?: Partial<Controllers>
-  userSupplier?: () => Express.User
-}): Express {
-  auth.default.authenticationMiddleware = () => (req, res, next) => next()
-  return appSetup(controllers as Controllers, production, userSupplier)
-}
